+++ StripeCore/StripeCoreTests/API Bindings/STPTelemetryClientFunctionalTest.swift
@@ -9,12 +9,13 @@
 import XCTest
 
 // swift-format-ignore
-@testable @_spi(STP) import StripeApplePay
-
-// swift-format-ignore
 @testable @_spi(STP) import StripeCore
 
 class STPTelemetryClientFunctionalTest: XCTestCase {
+    override func setUp() async throws {
+        STPTelemetryClient._forceShouldSendTelemetryInTests = true
+    }
+
     func testSendFraudDetectionData() {
         // Sending telemetry without any FraudDetectionData...
         FraudDetectionData.shared.sid = nil
@@ -22,7 +23,7 @@ class STPTelemetryClientFunctionalTest: XCTestCase {
         FraudDetectionData.shared.muid = nil
         FraudDetectionData.shared.guid = nil
         let sendTelemetry1 = expectation(description: "")
-        STPTelemetryClient.shared.sendTelemetryData(forceSend: true) { _ in
+        STPTelemetryClient.shared.sendTelemetryData { _ in
             sendTelemetry1.fulfill()
         }
         waitForExpectations(timeout: 10, handler: nil)
@@ -33,16 +34,19 @@ class STPTelemetryClientFunctionalTest: XCTestCase {
         XCTAssertNotNil(sid)
         XCTAssertNotNil(muid)
         XCTAssertNotNil(guid)
+        // ...and updates UserDefaults
+        XCTAssertEqual(UserDefaults.standard.fraudDetectionData, FraudDetectionData.shared)
 
         let sendTelemetry2 = expectation(description: "")
         // Sending telemetry again...
-        STPTelemetryClient.shared.sendTelemetryData(forceSend: true) { _ in
+        STPTelemetryClient.shared.sendTelemetryData { _ in
             sendTelemetry2.fulfill()
         }
         // ...gives the same FraudDetectionData
         XCTAssertEqual(FraudDetectionData.shared.sid, sid)
         XCTAssertEqual(FraudDetectionData.shared.muid, muid)
         XCTAssertEqual(FraudDetectionData.shared.guid, guid)
+        XCTAssertEqual(UserDefaults.standard.fraudDetectionData, FraudDetectionData.shared) // make sure user defaults is set too
         guard let sidCreationDate = FraudDetectionData.shared.sidCreationDate else {
             XCTFail()
             return
@@ -55,7 +59,7 @@ class STPTelemetryClientFunctionalTest: XCTestCase {
         FraudDetectionData.shared.sidCreationDate = Date(timeIntervalSinceNow: -999999)
         let sendTelemetry3 = expectation(description: "")
         // ...and sending telemetry again
-        STPTelemetryClient.shared.sendTelemetryData(forceSend: true) { _ in
+        STPTelemetryClient.shared.sendTelemetryData { _ in
             sendTelemetry3.fulfill()
         }
         waitForExpectations(timeout: 10, handler: nil)
@@ -63,5 +67,6 @@ class STPTelemetryClientFunctionalTest: XCTestCase {
         XCTAssertEqual(FraudDetectionData.shared.muid, muid)
         XCTAssertEqual(FraudDetectionData.shared.guid, guid)
         XCTAssertNotEqual(FraudDetectionData.shared.sid, sid)
+        XCTAssertEqual(UserDefaults.standard.fraudDetectionData, FraudDetectionData.shared) // make sure user defaults is set too
     }
 }