+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## X.Y.Z - changes pending release
+## X.Y.Z - changes pending release 
+### All
+* [Fixed] Improved telemetry used for [advanced fraud signals](https://docs.stripe.com/disputes/prevention/advanced-fraud-detection).
 
 ### PaymentSheet
 * [Changed] Custom Payment Methods is now generally available.